requirements.txt
asgiref==3.7.2
cffi==1.15.1
colorama==0.4.6
cryptography==41.0.3
Django==3.2.20
django-environ==0.10.0
djangorestframework==3.14.0
drf-yasg==1.21.7
environ==1.0
inflection==0.5.1
jwt==1.3.1
loguru==0.7.0
packaging==23.1
Pillow==10.0.0
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
win32-setctime==1.1.0
